backend/requirements.txt
# -- Core --
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic>=2.0
pydantic-settings>=2.0
python-multipart>=0.0.18

# -- Database --
sqlalchemy[asyncio]>=2.0
asyncpg>=0.30.0
psycopg2-binary>=2.9
alembic>=1.14.0

# -- Task queue --
celery[redis]>=5.4.0
redis>=5.0

# -- Auth --
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7

# -- Email --
aiosmtplib>=3.0

# -- HTTP client --
httpx>=0.28.0

# -- Reporting --
openpyxl>=3.1
jinja2>=3.1
weasyprint>=63.0

# -- HTML parsing (landing page cloner) --
beautifulsoup4>=4.12
lxml>=5.0

# -- Security hardening --
slowapi>=0.1.9
bleach>=6.0
cryptography>=42.0

# -- AI / Agent integration --
anthropic>=0.30.0
# mcp  # Model Context Protocol SDK -- optional, install when available

# -- Testing --
pytest>=8.0
pytest-asyncio>=0.25.0
pytest-cov>=6.0
factory-boy>=3.3
aiosqlite>=0.20.0
